completed task view, start styles
