+++ src/App.css
@@ -54,11 +54,42 @@ h1 {
 h2 {
   font-family: Verdana, Geneva, Tahoma, sans-serif;
   color: #ffffff;
+  font-size: 40px;
+}
+h3 {
+  font-family: Verdana, Geneva, Tahoma, sans-serif;
+  color: #52ab98;
+  font-size: 40px;
+}
+
+h5 {
+  font-family: Verdana, Geneva, Tahoma, sans-serif;
+  font-size: 25px;
 }
 
 li {
   list-style-type: none;
 
+}
+.taskEntry form input, select, textarea{
+  text-align: center;
+  font-family: 'Courier New', Courier, monospace;
+  font-size: 40px;
+  display: flex;
+  flex-direction: row;
+  margin-bottom: 20px;
+  
+}
+
+.taskEntry h3 {
+  text-align: center;
+}
+
+.taskEntry {
+
+  background-color: #c8d8e4;
+  padding-top: 50px;
+  padding-bottom: 2000px;
 }
 
 .links {
@@ -73,10 +104,41 @@ li {
   font-size: 30px;
 }
 
-.taskEntry {
+.taskForm {
+  display: flex;
+  justify-content: center;
+  
+  
+  
+}
+.taskHeader {
+  display: flex;
+}
+.taskHeader h5{
+  padding-left: 20px;
+  padding-right: 20px;
+
+}
+
+.taskFeed {
+  display: flex;
+  
+}
+.taskFeed p {
+  padding-right: 20px;
+  padding-left: 20px;
+  font-size: 20px;
+  
+}
+
+.entrySubmit {
+  font-size: 35px;
+  background-color:#c8d8e4 ;
+}
+
+.taskHolder {
   background-color: #c8d8e4;
-  padding-top: 200px;
+  padding-top: 50px;
   padding-bottom: 2000px;
   text-align: center;
-  
 }
