--- src/components/TaskMain.js
@@ -1,9 +1,0 @@
-const TaskMain = () => {
-    return (
-        <>
- 
-        </>
-    )
-}
-
-export default TaskMain;
